plant-disease-detector/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.1.0
Pillow>=9.3.0
Flask>=2.2.0
Flask-CORS>=3.0.10
Werkzeug>=2.2.0
opencv-python>=4.6.0
tqdm>=4.64.0
kaggle>=1.5.12
pytest>=7.2.0
gunicorn>=20.1.0
python-dotenv>=0.21.0
